refactor: 클래스를 세부 패키지로 분리
