+++ src/main/java/roomescape/model/Reservations.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.model;
 
 import java.util.ArrayList;
 import java.util.List;
