+++ src/test/java/roomescape/ReservationsTest.java
@@ -4,6 +4,9 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.model.Reservation;
+import roomescape.model.ReservationReqDto;
+import roomescape.model.Reservations;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
